# server.js
const express = require('express');
const app = express();
const PORT = 3000;
const swaggerUi = require('swagger-ui-express');
const openapiSpec = require('./openapi.json');

app.use('/docs', swaggerUi.serve, swaggerUi.setup(openapiSpec));


let tasks = [
  { id: 1, title: "Buy milk", done: false },
  { id: 2, title: "Walk the dog", done: true },
  { id: 3, title: "Write assignment", done: false }
];

app.use(express.json()); // it lets Express parse JSON request bodies

app.post('/tasks', (req, res) => {
  const { title } = req.body;

  if (!title || title.trim() === '') {
    return res.status(400).json({ error: "Title is required" });
  }

  const newTask = {
    id: tasks.length ? Math.max(...tasks.map(t => t.id)) + 1 : 1,
    title: title,
    done: false
  };

  tasks.push(newTask);
  res.status(201).json(newTask);
});

app.get('/tasks', (req, res) => {
  res.json(tasks);
});

app.put('/tasks/:id', (req, res) => {
  const task = tasks.find(t => t.id === Number(req.params.id));

  if (!task){
    return res.status(404).json({error: `Task ${req.params.id} is not found` })
  }

  const { title , done} = req.body;

  if(!title == undefined){
    task.title = title;
  }
  
  if (!done == undefined){
    task.done = done;
  }
  res.json(task);
  
});

app.delete('/tasks/:id', (req, res) => {
  const index = tasks.findIndex(t => t.id === Number(req.params.id));
  if (index === -1){
    res.status(404).json({error: `Task ${req.params.id} is not found`})
  }
  tasks.splice(index,1);
  res.status(204).send();
});


app.get('/tasks/:id', (req, res) => {
  const task = tasks.find(t => t.id === Number(req.params.id));
  if (!task) {
    return res.status(404).json({ error: `Task ${req.params.id} not found` });
  }
  res.json(task);
});

app.get('/' , (req,res) => {
    res.json({
        name: "Task API",
        version: "1.0",
        endpoints: ["/tasks"]
    });
});

app.get('/health', (req,res) => {
    res.json({status: "ok"});
});

app.listen(PORT , () => {
    console.log('Server running at http://localhost:${PORT}');
});